Fixed infinite loop.
Fixed an infinite loop and got the program to run through successfully.
I'll need to complete some more testing in the future but for now it seems to be working.

+++ cards.cpp
@@ -209,7 +209,7 @@ void Hand::add_card()
 	value += (double)(new_card.get_rank());
 }
 
-int Hand::get_value()
+double Hand::get_value()
 {
 	return value;
 }
